A-More OOP: Ui, Storage, Parser, TaskList
