+++ src/main/java/TaskList/Deadline.java
@@ -1,4 +1,5 @@
-package Sinep;
+package TaskList;
+
 public class Deadline extends Task {
     protected String dateTime;
     public Deadline(String description) {
