+++ src/main/java/TaskList/Event.java
@@ -1,4 +1,5 @@
-package Sinep;
+package TaskList;
+
 public class Event extends Task {
     protected String startDateTime;
     protected String endDateTime;
