+++ src/main/java/UI/BackboneUI.java
@@ -1,0 +1,6 @@
+package UI;
+
+public class BackboneUI {
+    public static String line = "_____________________________________________________________________";
+    public static String newLine = System.lineSeparator();
+}
